Zweather/requirements.txt
# Zedd Weather - Python Dependencies
# Node 1 (Telemetry), Node 2 (Orchestration), Edge InfluxDB Publisher,
# FastAPI backend, and Dash Python frontend.

# InfluxDB v2 time-series client (used by app.py / edge publisher)
influxdb-client>=1.50.0,<2.0

# MQTT
paho-mqtt>=1.6.1,<2.0

# Sense HAT
sense-hat>=2.6.0; platform_machine == "aarch64" or platform_machine == "armv7l"

# GPIO access (Raspberry Pi)
RPi.GPIO>=0.7.1; platform_machine == "aarch64" or platform_machine == "armv7l"
gpiozero>=2.0; platform_machine == "aarch64" or platform_machine == "armv7l"

# I2C / SMBus (for UV sensor and other I2C peripherals)
smbus2>=0.4.3

# Pimoroni Enviro+ sensors
enviroplus>=0.0.6; platform_machine == "aarch64" or platform_machine == "armv7l"
pimoroni-bme280>=1.0.0; platform_machine == "aarch64" or platform_machine == "armv7l"
pimoroni-ltr559>=1.0.0; platform_machine == "aarch64" or platform_machine == "armv7l"
pms5003>=1.0.0; platform_machine == "aarch64" or platform_machine == "armv7l"

# Modbus / RS485 industrial sensors
pymodbus>=3.6.0
pyserial>=3.5

# AI / Orchestration (Node 2)
google-genai>=1.0.0
aiohttp>=3.13.3
python-dotenv>=1.0.0

# REST API
fastapi>=0.115.0
uvicorn>=0.30.0

# HTTP client (used by Dash frontend and OllamaClient)
requests>=2.32.0

# ---------------------------------------------------------------------------
# Python Dash frontend
# ---------------------------------------------------------------------------
dash>=2.18.0
plotly>=5.24.0
dash-bootstrap-components>=1.6.0

# Raspberry Pi AI HAT+ (Hailo-8L NPU)
hailort>=4.18.0; platform_machine == "aarch64" or platform_machine == "armv7l"

# Testing
pytest>=8.0.0
pytest-mock>=3.0.0
